+++ validator/utils/logging.py
@@ -171,3 +171,188 @@ class TimeBasedLogger:
             self.last_log_time = current_time
             return True
         return False
+
+
+# Task preparation augmentation logging functions
+def _log_dpo_augmentations(augmentations: dict, dataset) -> None:
+    """Log DPO augmentation configuration details."""
+    import validator.core.constants as cst
+
+    logger = get_logger(__name__)
+    logger.info(
+        f"DPO Augmentations: rearrange={augmentations.get('rearrange_sentences')}, "
+        f"prompt_honeypot={augmentations.get('add_prompt_honeypot')}, "
+        f"response_honeypot={augmentations.get('add_response_honeypot')}, "
+        f"swap={augmentations.get('swap_chosen_rejected')}"
+    )
+
+    if augmentations.get("add_prompt_honeypot"):
+        logger.info(f"Prompt honeypot UID: {augmentations['prompt_uid']} (added to all prompts)")
+
+    if augmentations.get("add_response_honeypot"):
+        response_honeypot_indices = augmentations.get("response_honeypot_indices", set())
+        logger.info(
+            f"Response honeypot UID: {augmentations['response_uid']} "
+            f"(in {'chosen' if augmentations['honeypot_in_chosen'] else 'rejected'} "
+            f"at {'start' if augmentations['honeypot_at_start'] else 'end'} "
+            f"for {len(response_honeypot_indices)}/{len(dataset)} rows)"
+        )
+
+
+def _log_instruct_augmentations(augmentations: dict, dataset) -> None:
+    """Log Instruct augmentation configuration details."""
+    logger = get_logger(__name__)
+    logger.info(
+        f"Instruct Augmentations: rearrange_input={augmentations.get('rearrange_input')}, "
+        f"rearrange_output={augmentations.get('rearrange_output')}, "
+        f"input_honeypot={augmentations.get('add_input_honeypot')}, "
+        f"output_honeypot={augmentations.get('add_output_honeypot')}"
+    )
+
+    # Log word honeypot transformations
+    input_word_transforms = augmentations.get('input_apply_word_transforms', False)
+    output_word_transforms = augmentations.get('output_apply_word_transforms', False)
+    if input_word_transforms or output_word_transforms:
+        logger.info(f"Word Honeypots: input_transforms={input_word_transforms}, output_transforms={output_word_transforms}")
+
+        if input_word_transforms:
+            input_transform_type = augmentations.get('input_text_transform_type', 'none')
+            logger.info(f"  Input word transform: {input_transform_type}")
+
+        if output_word_transforms:
+            output_transform_type = augmentations.get('output_text_transform_type', 'none')
+            logger.info(f"  Output word transform: {output_transform_type}")
+
+    if augmentations.get("add_input_honeypot"):
+        logger.info(
+            f"Input honeypot UID: {augmentations['input_uid']} "
+            f"(at {'start' if augmentations.get('input_honeypot_at_start') else 'end'} of all instructions)"
+        )
+
+    if augmentations.get("add_output_honeypot"):
+        output_honeypot_indices = augmentations.get("output_honeypot_indices", set())
+        logger.info(
+            f"Output honeypot UID: {augmentations['output_uid']} "
+            f"(at {'start' if augmentations.get('output_honeypot_at_start') else 'end'} "
+            f"for {len(output_honeypot_indices)}/{len(dataset)} rows)"
+        )
+
+
+def _log_grpo_augmentations(augmentations: dict) -> None:
+    """Log GRPO augmentation configuration details."""
+    logger = get_logger(__name__)
+    logger.info("GRPO Augmentations: applying honeypots to prompts only")
+
+    if augmentations.get("add_prompt_honeypot"):
+        logger.info(f"  Prompt UID honeypot: {augmentations['prompt_uid']} (added to all prompts)")
+
+    if augmentations.get('apply_word_transforms'):
+        transform_type = augmentations.get('transform_type', 'none')
+        logger.info(f"  GRPO word transforms: {transform_type}")
+
+
+def _log_dpo_examples(result: list[dict]) -> None:
+    """Log examples of augmented DPO data."""
+    import validator.core.constants as cst
+
+    logger = get_logger(__name__)
+    logger.info("[DPO_AUGMENTATION] Showing 2 examples of augmented data:")
+
+    # Show first 2 examples to see the augmentations
+    for i in range(min(2, len(result))):
+        example = result[i]
+        logger.info(f"[DPO_EXAMPLE_{i + 1}]:")
+
+        # Show prompt (truncate if too long)
+        prompt = example.get(cst.STANDARD_DPO_PROMPT_COLUMN, "")
+        if len(prompt) > 150:
+            prompt_preview = prompt[:150] + "..."
+        else:
+            prompt_preview = prompt
+        logger.info(f"  Prompt: {prompt_preview}")
+
+        # Show chosen (truncate if too long)
+        chosen = example.get(cst.STANDARD_DPO_CHOSEN_COLUMN, "")
+        if len(chosen) > 150:
+            chosen_preview = chosen[:150] + "..."
+        else:
+            chosen_preview = chosen
+        logger.info(f"  Chosen: {chosen_preview}")
+
+        # Show rejected (truncate if too long)
+        rejected = example.get(cst.STANDARD_DPO_REJECTED_COLUMN, "")
+        if len(rejected) > 150:
+            rejected_preview = rejected[:150] + "..."
+        else:
+            rejected_preview = rejected
+        logger.info(f"  Rejected: {rejected_preview}")
+
+
+def _log_instruct_examples(result: list[dict]) -> None:
+    """Log examples of augmented Instruct data."""
+    import validator.core.constants as cst
+
+    logger = get_logger(__name__)
+    logger.info("[INSTRUCT_AUGMENTATION] Showing 2 examples of augmented data:")
+
+    # Show first 2 examples to see the augmentations
+    for i in range(min(2, len(result))):
+        example = result[i]
+        logger.info(f"[INSTRUCT_EXAMPLE_{i + 1}]:")
+
+        # Show instruction (truncate if too long)
+        instruction = example.get(cst.STANDARD_INSTRUCT_COLUMN, "")
+        if len(instruction) > 150:
+            instruction_preview = instruction[:150] + "..."
+        else:
+            instruction_preview = instruction
+        logger.info(f"  Instruction: {instruction_preview}")
+
+        # Show output (truncate if too long)
+        output = example.get(cst.STANDARD_OUTPUT_COLUMN, "")
+        if len(output) > 150:
+            output_preview = output[:150] + "..."
+        else:
+            output_preview = output
+        logger.info(f"  Output: {output_preview}")
+
+        # Show input if present
+        if cst.STANDARD_INPUT_COLUMN in example:
+            input_text = example.get(cst.STANDARD_INPUT_COLUMN, "")
+            if input_text and len(input_text) > 150:
+                input_preview = input_text[:150] + "..."
+            else:
+                input_preview = input_text
+            if input_preview:
+                logger.info(f"  Input: {input_preview}")
+
+
+def _log_grpo_examples(result: list[dict]) -> None:
+    """Log examples of augmented GRPO data."""
+    import validator.core.constants as cst
+
+    logger = get_logger(__name__)
+    logger.info("[GRPO_AUGMENTATION] Showing 2 examples of augmented data:")
+
+    # Show first 2 examples to see the augmentations
+    for i in range(min(2, len(result))):
+        example = result[i]
+        logger.info(f"[GRPO_EXAMPLE_{i + 1}]:")
+
+        # Show prompt (truncate if too long)
+        prompt = example.get(cst.STANDARD_GRPO_PROMPT_COLUMN, "")
+        if len(prompt) > 150:
+            prompt_preview = prompt[:150] + "..."
+        else:
+            prompt_preview = prompt
+        logger.info(f"  Prompt: {prompt_preview}")
+
+        # Show extra column if present
+        if cst.STANDARD_GRPO_EXTRA_COLUMN in example:
+            extra = example.get(cst.STANDARD_GRPO_EXTRA_COLUMN, "")
+            if extra and len(extra) > 150:
+                extra_preview = extra[:150] + "..."
+            else:
+                extra_preview = extra
+            if extra_preview:
+                logger.info(f"  Extra: {extra_preview}")
